chore(op-challenger): address comments(remove unused filter)
