+++ op-challenger/game/fault/contracts/delayed_weth.go
@@ -32,7 +32,7 @@ func NewDelayedWETHContract(metrics metrics.ContractMetricer, addr common.Addres
 	return &DelayedWETHContract{
 		metrics:     metrics,
 		multiCaller: caller,
-		contract:    batching.NewBoundContract(contractAbi, addr, caller),
+		contract:    batching.NewBoundContract(contractAbi, addr),
 	}
 }
