+++ op-challenger/game/fault/contracts/faultdisputegame.go
@@ -13,7 +13,6 @@ import (
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/contracts/metrics"
 	"github.com/ethereum-optimism/optimism/op-challenger/game/fault/types"
 	gameTypes "github.com/ethereum-optimism/optimism/op-challenger/game/types"
-	"github.com/ethereum-optimism/optimism/op-e2e/bindings"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching"
 	"github.com/ethereum-optimism/optimism/op-service/sources/batching/rpcblock"
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
@@ -98,7 +97,7 @@ func NewFaultDisputeGameContract(ctx context.Context, metrics metrics.ContractMe
 			FaultDisputeGameContractLatest: FaultDisputeGameContractLatest{
 				metrics:     metrics,
 				multiCaller: caller,
-				contract:    batching.NewBoundContract(legacyAbi, addr, caller),
+				contract:    batching.NewBoundContract(legacyAbi, addr),
 			},
 		}, nil
 	} else if strings.HasPrefix(version, "0.18.") || strings.HasPrefix(version, "1.0.") {
@@ -108,7 +107,7 @@ func NewFaultDisputeGameContract(ctx context.Context, metrics metrics.ContractMe
 			FaultDisputeGameContractLatest: FaultDisputeGameContractLatest{
 				metrics:     metrics,
 				multiCaller: caller,
-				contract:    batching.NewBoundContract(legacyAbi, addr, caller),
+				contract:    batching.NewBoundContract(legacyAbi, addr),
 			},
 		}, nil
 	} else if strings.HasPrefix(version, "1.1.") {
@@ -118,14 +117,14 @@ func NewFaultDisputeGameContract(ctx context.Context, metrics metrics.ContractMe
 			FaultDisputeGameContractLatest: FaultDisputeGameContractLatest{
 				metrics:     metrics,
 				multiCaller: caller,
-				contract:    batching.NewBoundContract(legacyAbi, addr, caller),
+				contract:    batching.NewBoundContract(legacyAbi, addr),
 			},
 		}, nil
 	} else {
 		return &FaultDisputeGameContractLatest{
 			metrics:     metrics,
 			multiCaller: caller,
-			contract:    batching.NewBoundContract(contractAbi, addr, caller),
+			contract:    batching.NewBoundContract(contractAbi, addr),
 		}, nil
 	}
 }
@@ -442,57 +441,6 @@ func (f *FaultDisputeGameContractLatest) GetAllClaims(ctx context.Context, block
 		claims = append(claims, f.decodeClaim(result, idx))
 	}
 	return claims, nil
-}
-
-func (f *FaultDisputeGameContractLatest) GetSubClaims(ctx context.Context, block rpcblock.Block, aggClaim *types.Claim) ([]common.Hash, error) {
-	defer f.metrics.StartContractRequest("GetAllSubClaims")()
-
-	filter, err := bindings.NewFaultDisputeGameFilterer(f.contract.Addr(), f.multiCaller)
-	if err != nil {
-		return nil, err
-	}
-
-	parentIndex := [...]*big.Int{big.NewInt(int64(aggClaim.ParentContractIndex))}
-	claim := [...][32]byte{aggClaim.ClaimData.ValueBytes()}
-	claimant := [...]common.Address{aggClaim.Claimant}
-	moveIter, err := filter.FilterMove(nil, parentIndex[:], claim[:], claimant[:])
-	if err != nil {
-		return nil, fmt.Errorf("failed to filter move event log: %w", err)
-	}
-	ok := moveIter.Next()
-	if !ok {
-		return nil, fmt.Errorf("failed to get move event log: %w", moveIter.Error())
-	}
-	txHash := moveIter.Event.Raw.TxHash
-
-	// todo: replace hardcoded method name
-	txCall := batching.NewTxGetByHash(f.contract.Abi(), txHash, "move")
-	result, err := f.multiCaller.SingleCall(ctx, rpcblock.Latest, txCall)
-	if err != nil {
-		return nil, fmt.Errorf("failed to load claim calldata: %w", err)
-	}
-
-	txn, err := txCall.DecodeToTx(result)
-	if err != nil {
-		return nil, fmt.Errorf("failed to decode tx: %w", err)
-	}
-
-	var subClaims []common.Hash
-
-	if len(txn.BlobHashes()) > 0 {
-		// todo: fetch Blobs and unpack it into subClaims
-		return nil, fmt.Errorf("blob tx hasn't been supported")
-	} else {
-		inputMap, err := txCall.UnpackCallData(txn)
-		if err != nil {
-			return nil, fmt.Errorf("failed to unpack tx resp: %w", err)
-		}
-		// todo: replace claim with nary-subclaims
-		claim := *abi.ConvertType(inputMap[subClaimField], new([32]byte)).(*[32]byte)
-		subClaims = append(subClaims, claim)
-	}
-
-	return subClaims, nil
 }
 
 func (f *FaultDisputeGameContractLatest) IsResolved(ctx context.Context, block rpcblock.Block, claims ...types.Claim) ([]bool, error) {
